refactor(chart.commands): separate commands into two different modules refactor: move imports under TYPE_CHECKING
