+++ superset/charts/schemas.py
@@ -26,7 +26,6 @@ from marshmallow_enum import EnumField
 
 from superset import app
 from superset.common.chart_data import ChartDataResultFormat, ChartDataResultType
-from superset.common.query_context_factory import QueryContextFactory
 from superset.db_engine_specs.base import builtin_time_grains
 from superset.utils import schema as utils
 from superset.utils.core import (
@@ -39,6 +38,7 @@ from superset.utils.core import (
 
 if TYPE_CHECKING:
     from superset.common.query_context import QueryContext
+    from superset.common.query_context_factory import QueryContextFactory
 
 config = app.config
 
@@ -1153,6 +1153,9 @@ class ChartDataQueryContextSchema(Schema):
 
     def get_query_context_factory(self) -> QueryContextFactory:
         if self.query_context_factory is None:
+            # pylint: disable=import-outside-toplevel
+            from superset.common.query_context_factory import QueryContextFactory
+
             self.query_context_factory = QueryContextFactory()
         return self.query_context_factory
