+++ tests/integration_tests/tasks/async_queries_tests.py
@@ -23,7 +23,7 @@ from celery.exceptions import SoftTimeLimitExceeded
 from flask import g
 
 from superset.charts.commands.exceptions import ChartDataQueryFailedError
-from superset.charts.data.commands import ChartDataCommand
+from superset.charts.data.commands.get_data_command import ChartDataCommand
 from superset.exceptions import SupersetException
 from superset.extensions import async_query_manager, security_manager
 from superset.tasks import async_queries
